REFACTOR (src/components/NavBar/MobileMenu/index.tsx) (src/app/layout.tsx) (package-lock.json) Coponetiza navbar web e navbar mobile @thayanneregoo @stephanybrazeir0
